pass logger from the factory
