+++ ipc/interface.go
@@ -1,4 +1,4 @@
-package transport
+package ipc
 
 import (
 	"context"
