+++ ipc/pipe/pipe_factory_test.go
@@ -1,6 +1,7 @@
 package pipe
 
 import (
+	"context"
 	"os/exec"
 	"sync"
 	"testing"
@@ -13,10 +14,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_GetState2(t *testing.T) {
+func Test_GetState(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 		assert.Equal(t, worker.StateStopped, w.State().Value())
@@ -26,13 +29,18 @@ func Test_GetState2(t *testing.T) {
 	assert.NotNil(t, w)
 
 	assert.Equal(t, worker.StateReady, w.State().Value())
-	assert.NoError(t, w.Stop())
+	err = w.Stop()
+	if err != nil {
+		t.Errorf("error stopping the Process: error %v", err)
+	}
 }
 
-func Test_Kill2(t *testing.T) {
+func Test_Kill(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	wg := &sync.WaitGroup{}
 	wg.Add(1)
 	go func() {
@@ -52,10 +60,12 @@ func Test_Kill2(t *testing.T) {
 	wg.Wait()
 }
 
-func Test_Pipe_Start2(t *testing.T) {
+func Test_Pipe_Start(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -66,64 +76,86 @@ func Test_Pipe_Start2(t *testing.T) {
 	assert.NoError(t, w.Stop())
 }
 
-func Test_Pipe_StartError2(t *testing.T) {
+func Test_Pipe_StartError(t *testing.T) {
+	t.Parallel()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 	err := cmd.Start()
 	if err != nil {
 		t.Errorf("error running the command: error %v", err)
 	}
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Pipe_PipeError3(t *testing.T) {
+func Test_Pipe_PipeError(t *testing.T) {
+	t.Parallel()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 	_, err := cmd.StdinPipe()
 	if err != nil {
 		t.Errorf("error creating the STDIN pipe: error %v", err)
 	}
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Pipe_PipeError4(t *testing.T) {
+func Test_Pipe_PipeError2(t *testing.T) {
+	t.Parallel()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
+	// error cause
 	_, err := cmd.StdinPipe()
 	if err != nil {
 		t.Errorf("error creating the STDIN pipe: error %v", err)
 	}
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Pipe_Failboot2(t *testing.T) {
+func Test_Pipe_Failboot(t *testing.T) {
 	cmd := exec.Command("php", "../../tests/failboot.php")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
+
 	assert.Nil(t, w)
 	assert.Error(t, err)
 }
 
-func Test_Pipe_Invalid2(t *testing.T) {
+func Test_Pipe_Invalid(t *testing.T) {
+	t.Parallel()
 	cmd := exec.Command("php", "../../tests/invalid.php")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Pipe_Echo2(t *testing.T) {
+func Test_Pipe_Echo(t *testing.T) {
+	t.Parallel()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
-	assert.NoError(t, err)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	defer func() {
+		err = w.Stop()
+		if err != nil {
+			t.Errorf("error stopping the Process: error %v", err)
+		}
+	}()
 
 	sw := worker.From(w)
 	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
-
 	assert.NoError(t, err)
 	assert.NotNil(t, res)
 	assert.NotNil(t, res.Body)
@@ -136,15 +168,51 @@ func Test_Pipe_Echo2(t *testing.T) {
 	}()
 
 	assert.Equal(t, "hello", res.String())
-	err = w.Stop()
-	assert.NoError(t, err)
 }
 
-func Test_Pipe_Broken2(t *testing.T) {
-	cmd := exec.Command("php", "../../tests/client.php", "broken", "pipes")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+func Test_Pipe_Echo_Script(t *testing.T) {
+	t.Parallel()
+	cmd := exec.Command("sh", "../../tests/pipes_test_script.sh")
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer func() {
+		err = w.Stop()
+		if err != nil {
+			t.Errorf("error stopping the Process: error %v", err)
+		}
+	}()
+
+	sw := worker.From(w)
+	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
 	assert.NoError(t, err)
+	assert.NotNil(t, res)
+	assert.NotNil(t, res.Body)
+	assert.Empty(t, res.Context)
+
+	go func() {
+		if w.Wait() != nil {
+			t.Fail()
+		}
+	}()
+
+	assert.Equal(t, "hello", res.String())
+}
+
+func Test_Pipe_Broken(t *testing.T) {
+	t.Parallel()
+	cmd := exec.Command("php", "../../tests/client.php", "broken", "pipes")
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
+	require.NoError(t, err)
 	require.NotNil(t, w)
+
+	go func() {
+		errW := w.Wait()
+		require.Error(t, errW)
+	}()
 
 	sw := worker.From(w)
 	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
@@ -153,14 +221,14 @@ func Test_Pipe_Broken2(t *testing.T) {
 
 	time.Sleep(time.Second)
 	err = w.Stop()
-	assert.Error(t, err)
+	assert.NoError(t, err)
 }
 
-func Benchmark_Pipe_SpawnWorker_Stop2(b *testing.B) {
-	f := NewPipeFactory()
+func Benchmark_Pipe_SpawnWorker_Stop(b *testing.B) {
+	f := NewPipeFactory(log)
 	for n := 0; n < b.N; n++ {
 		cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
-		w, _ := f.SpawnWorker(cmd)
+		w, _ := f.SpawnWorkerWithTimeout(context.Background(), cmd)
 		go func() {
 			if w.Wait() != nil {
 				b.Fail()
@@ -174,10 +242,10 @@ func Benchmark_Pipe_SpawnWorker_Stop2(b *testing.B) {
 	}
 }
 
-func Benchmark_Pipe_Worker_ExecEcho2(b *testing.B) {
+func Benchmark_Pipe_Worker_ExecEcho(b *testing.B) {
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(context.Background(), cmd)
 	sw := worker.From(w)
 
 	b.ReportAllocs()
@@ -202,9 +270,10 @@ func Benchmark_Pipe_Worker_ExecEcho2(b *testing.B) {
 	}
 }
 
-func Benchmark_Pipe_Worker_ExecEcho4(b *testing.B) {
+func Benchmark_Pipe_Worker_ExecEcho3(b *testing.B) {
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -225,9 +294,10 @@ func Benchmark_Pipe_Worker_ExecEcho4(b *testing.B) {
 	}
 }
 
-func Benchmark_Pipe_Worker_ExecEchoWithoutContext2(b *testing.B) {
+func Benchmark_Pipe_Worker_ExecEchoWithoutContext(b *testing.B) {
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	ctx := context.Background()
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -248,16 +318,17 @@ func Benchmark_Pipe_Worker_ExecEchoWithoutContext2(b *testing.B) {
 	}
 }
 
-func Test_Echo2(t *testing.T) {
+func Test_Echo(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	sw := worker.From(w)
-
 	go func() {
 		assert.NoError(t, sw.Wait())
 	}()
@@ -278,10 +349,12 @@ func Test_Echo2(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Test_BadPayload2(t *testing.T) {
+func Test_BadPayload(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 
 	sw := worker.From(w)
 
@@ -296,16 +369,19 @@ func Test_BadPayload2(t *testing.T) {
 	}()
 
 	res, err := sw.Exec(&payload.Payload{})
+
 	assert.Error(t, err)
 	assert.Nil(t, res)
 
 	assert.Contains(t, err.Error(), "payload can not be empty")
 }
 
-func Test_String2(t *testing.T) {
+func Test_String(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
@@ -321,10 +397,12 @@ func Test_String2(t *testing.T) {
 	assert.Contains(t, w.String(), "num_execs: 0")
 }
 
-func Test_Echo_Slow2(t *testing.T) {
+func Test_Echo_Slow(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/slow-client.php", "echo", "pipes", "10", "10")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
@@ -347,15 +425,17 @@ func Test_Echo_Slow2(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Test_Broken2(t *testing.T) {
+func Test_Broken(t *testing.T) {
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "broken", "pipes")
 
-	w, err := NewPipeFactory().SpawnWorker(cmd)
+	w, err := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	sw := worker.From(w)
+
 	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
 	assert.NotNil(t, err)
 	assert.Nil(t, res)
@@ -364,10 +444,12 @@ func Test_Broken2(t *testing.T) {
 	assert.Error(t, w.Stop())
 }
 
-func Test_Error2(t *testing.T) {
+func Test_Error(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "error", "pipes")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
@@ -391,10 +473,12 @@ func Test_Error2(t *testing.T) {
 	assert.Contains(t, err.Error(), "hello")
 }
 
-func Test_NumExecs2(t *testing.T) {
+func Test_NumExecs(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "pipes")
 
-	w, _ := NewPipeFactory().SpawnWorker(cmd)
+	w, _ := NewPipeFactory(log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
