+++ ipc/socket/socket_factory_spawn_test.go
@@ -1,10 +1,10 @@
 package socket
 
 import (
-	"context"
 	"net"
 	"os/exec"
 	"sync"
+	"syscall"
 	"testing"
 	"time"
 
@@ -12,17 +12,17 @@ import (
 	"github.com/spiral/roadrunner/v2/worker"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
 )
 
-func Test_Tcp_Start(t *testing.T) {
-	ctx := context.Background()
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
+var log = zap.NewNop()
 
+func Test_Tcp_Start2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -32,7 +32,7 @@ func Test_Tcp_Start(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -46,9 +46,7 @@ func Test_Tcp_Start(t *testing.T) {
 	}
 }
 
-func Test_Tcp_StartCloseFactory(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
+func Test_Tcp_StartCloseFactory2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 	} else {
@@ -56,7 +54,7 @@ func Test_Tcp_StartCloseFactory(t *testing.T) {
 	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	defer func() {
 		err = ls.Close()
 		if err != nil {
@@ -64,7 +62,7 @@ func Test_Tcp_StartCloseFactory(t *testing.T) {
 		}
 	}()
 
-	w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := f.SpawnWorker(cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -78,16 +76,12 @@ func Test_Tcp_StartCloseFactory(t *testing.T) {
 	}
 }
 
-func Test_Tcp_StartError(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
+func Test_Tcp_StartError2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -101,17 +95,12 @@ func Test_Tcp_StartError(t *testing.T) {
 		t.Errorf("error executing the command: error %v", err)
 	}
 
-	serv := NewSocketServer(ls, time.Minute)
-	time.Sleep(time.Second * 2)
-	w, err := serv.SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Tcp_Failboot(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
-
+func Test_Tcp_Failboot2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
@@ -126,42 +115,17 @@ func Test_Tcp_Failboot(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/failboot.php")
 
-	w, err2 := NewSocketServer(ls, time.Second*5).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err2 := NewSocketServer(ls, time.Second*5, log).SpawnWorker(cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err2)
 }
 
-func Test_Tcp_Timeout(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
+func Test_Tcp_Invalid2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
-
-	cmd := exec.Command("php", "../../tests/slow-client.php", "echo", "tcp", "200", "0")
-
-	w, err := NewSocketServer(ls, time.Millisecond*1).SpawnWorkerWithTimeout(ctx, cmd)
-	assert.Nil(t, w)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "context deadline exceeded")
-}
-
-func Test_Tcp_Invalid(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
-	ls, err := net.Listen("tcp", "127.0.0.1:9007")
-	if assert.NoError(t, err) {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -171,14 +135,12 @@ func Test_Tcp_Invalid(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/invalid.php")
 
-	w, err := NewSocketServer(ls, time.Second*1).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Second*1, log).SpawnWorker(cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Tcp_Broken(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
+func Test_Tcp_Broken2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
@@ -193,7 +155,7 @@ func Test_Tcp_Broken(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "broken", "tcp")
 
-	w, err := NewSocketServer(ls, time.Second*10).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -214,18 +176,16 @@ func Test_Tcp_Broken(t *testing.T) {
 	time.Sleep(time.Second)
 	err2 := w.Stop()
 	// write tcp 127.0.0.1:9007->127.0.0.1:34204: use of closed network connection
-	// but process is stopped
+	// but process exited
 	assert.NoError(t, err2)
 }
 
-func Test_Tcp_Echo(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
+func Test_Tcp_Echo2(t *testing.T) {
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -235,7 +195,7 @@ func Test_Tcp_Echo(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, _ := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, _ := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
@@ -258,63 +218,17 @@ func Test_Tcp_Echo(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Test_Tcp_Echo_Script(t *testing.T) {
-	time.Sleep(time.Millisecond * 10) // to ensure free socket
-	ctx := context.Background()
-	ls, err := net.Listen("tcp", "127.0.0.1:9007")
-	if assert.NoError(t, err) {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
-
-	cmd := exec.Command("sh", "../../tests/socket_test_script.sh")
-
-	w, _ := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
-	go func() {
-		assert.NoError(t, w.Wait())
-	}()
-	defer func() {
-		err = w.Stop()
-		if err != nil {
-			t.Errorf("error stopping the Process: error %v", err)
-		}
-	}()
-
-	sw := worker.From(w)
-
-	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
-
-	assert.NoError(t, err)
-	assert.NotNil(t, res)
-	assert.NotNil(t, res.Body)
-	assert.Empty(t, res.Context)
-
-	assert.Equal(t, "hello", res.String())
-}
-
-func Test_Unix_Start(t *testing.T) {
-	ctx := context.Background()
+func Test_Unix_Start2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(t, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -328,86 +242,63 @@ func Test_Unix_Start(t *testing.T) {
 	}
 }
 
-func Test_Unix_Failboot(t *testing.T) {
+func Test_Unix_Failboot2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	ctx := context.Background()
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(t, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/failboot.php")
 
-	w, err := NewSocketServer(ls, time.Second*5).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Second*5, log).SpawnWorker(cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err)
 }
 
-func Test_Unix_Timeout(t *testing.T) {
+func Test_Unix_Timeout2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	ctx := context.Background()
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(t, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/slow-client.php", "echo", "unix", "200", "0")
 
-	w, err := NewSocketServer(ls, time.Millisecond*100).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Millisecond*100, log).SpawnWorker(cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "context deadline exceeded")
+	assert.Contains(t, err.Error(), "relay timeout")
 }
 
-func Test_Unix_Invalid(t *testing.T) {
-	ctx := context.Background()
+func Test_Unix_Invalid2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(t, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/invalid.php")
 
-	w, err := NewSocketServer(ls, time.Second*10).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Second*10, log).SpawnWorker(cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Unix_Broken(t *testing.T) {
-	ctx := context.Background()
+func Test_Unix_Broken2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	if err == nil {
-		defer func() {
-			errC := ls.Close()
-			if errC != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		errC := ls.Close()
+		assert.NoError(t, errC)
+	}()
 
 	cmd := exec.Command("php", "../../tests/client.php", "broken", "unix")
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -424,31 +315,24 @@ func Test_Unix_Broken(t *testing.T) {
 
 	assert.Error(t, err)
 	assert.Nil(t, res)
+	wg.Wait()
 
 	time.Sleep(time.Second)
 	err = w.Stop()
 	assert.NoError(t, err)
-
-	wg.Wait()
 }
 
-func Test_Unix_Echo(t *testing.T) {
-	ctx := context.Background()
+func Test_Unix_Echo2(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				t.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		t.Skip("socket is busy")
-	}
+	assert.NoError(t, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(t, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -474,25 +358,19 @@ func Test_Unix_Echo(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Benchmark_Tcp_SpawnWorker_Stop(b *testing.B) {
-	ctx := context.Background()
-	ls, err := net.Listen("tcp", "127.0.0.1:9007")
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				b.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		b.Skip("socket is busy")
-	}
+func Benchmark_Tcp_SpawnWorker_Stop2(b *testing.B) {
+	ls, err := net.Listen("unix", "sock.unix")
+	assert.NoError(b, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(b, err)
+	}()
 
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	for n := 0; n < b.N; n++ {
 		cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-		w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
+		w, err := f.SpawnWorker(cmd)
 		if err != nil {
 			b.Fatal(err)
 		}
@@ -507,23 +385,17 @@ func Benchmark_Tcp_SpawnWorker_Stop(b *testing.B) {
 	}
 }
 
-func Benchmark_Tcp_Worker_ExecEcho(b *testing.B) {
-	ctx := context.Background()
-	ls, err := net.Listen("tcp", "127.0.0.1:9007")
-	if err == nil {
-		defer func() {
-			err = ls.Close()
-			if err != nil {
-				b.Errorf("error closing the listener: error %v", err)
-			}
-		}()
-	} else {
-		b.Skip("socket is busy")
-	}
+func Benchmark_Tcp_Worker_ExecEcho2(b *testing.B) {
+	ls, err := net.Listen("unix", "sock.unix")
+	assert.NoError(b, err)
+	defer func() {
+		err = ls.Close()
+		assert.NoError(b, err)
+	}()
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -543,13 +415,15 @@ func Benchmark_Tcp_Worker_ExecEcho(b *testing.B) {
 	}
 }
 
-func Benchmark_Unix_SpawnWorker_Stop(b *testing.B) {
-	ctx := context.Background()
+func Benchmark_Unix_SpawnWorker_Stop2(b *testing.B) {
+	defer func() {
+		_ = syscall.Unlink("sock.unix")
+	}()
 	ls, err := net.Listen("unix", "sock.unix")
 	if err == nil {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				b.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -557,11 +431,11 @@ func Benchmark_Unix_SpawnWorker_Stop(b *testing.B) {
 		b.Skip("socket is busy")
 	}
 
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	for n := 0; n < b.N; n++ {
 		cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-		w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
+		w, err := f.SpawnWorker(cmd)
 		if err != nil {
 			b.Fatal(err)
 		}
@@ -572,13 +446,15 @@ func Benchmark_Unix_SpawnWorker_Stop(b *testing.B) {
 	}
 }
 
-func Benchmark_Unix_Worker_ExecEcho(b *testing.B) {
-	ctx := context.Background()
+func Benchmark_Unix_Worker_ExecEcho2(b *testing.B) {
+	defer func() {
+		_ = syscall.Unlink("sock.unix")
+	}()
 	ls, err := net.Listen("unix", "sock.unix")
 	if err == nil {
 		defer func() {
-			err = ls.Close()
-			if err != nil {
+			errC := ls.Close()
+			if errC != nil {
 				b.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -588,7 +464,7 @@ func Benchmark_Unix_Worker_ExecEcho(b *testing.B) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorkerWithTimeout(ctx, cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorker(cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
