+++ ipc/socket/socket_factory_test.go
@@ -1,10 +1,10 @@
 package socket
 
 import (
+	"context"
 	"net"
 	"os/exec"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 
@@ -14,12 +14,15 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_Tcp_Start2(t *testing.T) {
+func Test_Tcp_Start(t *testing.T) {
+	ctx := context.Background()
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -29,7 +32,7 @@ func Test_Tcp_Start2(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -43,7 +46,9 @@ func Test_Tcp_Start2(t *testing.T) {
 	}
 }
 
-func Test_Tcp_StartCloseFactory2(t *testing.T) {
+func Test_Tcp_StartCloseFactory(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 	} else {
@@ -51,7 +56,7 @@ func Test_Tcp_StartCloseFactory2(t *testing.T) {
 	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	defer func() {
 		err = ls.Close()
 		if err != nil {
@@ -59,7 +64,7 @@ func Test_Tcp_StartCloseFactory2(t *testing.T) {
 		}
 	}()
 
-	w, err := f.SpawnWorker(cmd)
+	w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -73,12 +78,16 @@ func Test_Tcp_StartCloseFactory2(t *testing.T) {
 	}
 }
 
-func Test_Tcp_StartError2(t *testing.T) {
+func Test_Tcp_StartError(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -92,12 +101,17 @@ func Test_Tcp_StartError2(t *testing.T) {
 		t.Errorf("error executing the command: error %v", err)
 	}
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	serv := NewSocketServer(ls, time.Minute, log)
+	time.Sleep(time.Second * 2)
+	w, err := serv.SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Tcp_Failboot2(t *testing.T) {
+func Test_Tcp_Failboot(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
+
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
@@ -112,17 +126,42 @@ func Test_Tcp_Failboot2(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/failboot.php")
 
-	w, err2 := NewSocketServer(ls, time.Second*5).SpawnWorker(cmd)
+	w, err2 := NewSocketServer(ls, time.Second*5, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err2)
 }
 
-func Test_Tcp_Invalid2(t *testing.T) {
+func Test_Tcp_Timeout(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
+
+	cmd := exec.Command("php", "../../tests/slow-client.php", "echo", "tcp", "200", "0")
+
+	w, err := NewSocketServer(ls, time.Millisecond*1, log).SpawnWorkerWithTimeout(ctx, cmd)
+	assert.Nil(t, w)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "context deadline exceeded")
+}
+
+func Test_Tcp_Invalid(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
+	ls, err := net.Listen("tcp", "127.0.0.1:9007")
+	if assert.NoError(t, err) {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -132,12 +171,14 @@ func Test_Tcp_Invalid2(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/invalid.php")
 
-	w, err := NewSocketServer(ls, time.Second*1).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Second*1, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Tcp_Broken2(t *testing.T) {
+func Test_Tcp_Broken(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
@@ -152,7 +193,7 @@ func Test_Tcp_Broken2(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "broken", "tcp")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Second*10, log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -173,16 +214,18 @@ func Test_Tcp_Broken2(t *testing.T) {
 	time.Sleep(time.Second)
 	err2 := w.Stop()
 	// write tcp 127.0.0.1:9007->127.0.0.1:34204: use of closed network connection
-	// but process exited
+	// but process is stopped
 	assert.NoError(t, err2)
 }
 
-func Test_Tcp_Echo2(t *testing.T) {
+func Test_Tcp_Echo(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
 	ls, err := net.Listen("tcp", "127.0.0.1:9007")
 	if assert.NoError(t, err) {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
 				t.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -192,7 +235,7 @@ func Test_Tcp_Echo2(t *testing.T) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, _ := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, _ := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	go func() {
 		assert.NoError(t, w.Wait())
 	}()
@@ -215,17 +258,63 @@ func Test_Tcp_Echo2(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Test_Unix_Start2(t *testing.T) {
-	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(t, err)
+func Test_Tcp_Echo_Script(t *testing.T) {
+	time.Sleep(time.Millisecond * 10) // to ensure free socket
+	ctx := context.Background()
+	ls, err := net.Listen("tcp", "127.0.0.1:9007")
+	if assert.NoError(t, err) {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
+
+	cmd := exec.Command("sh", "../../tests/socket_test_script.sh")
+
+	w, _ := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
+	go func() {
+		assert.NoError(t, w.Wait())
 	}()
+	defer func() {
+		err = w.Stop()
+		if err != nil {
+			t.Errorf("error stopping the Process: error %v", err)
+		}
+	}()
+
+	sw := worker.From(w)
+
+	res, err := sw.Exec(&payload.Payload{Body: []byte("hello")})
+
+	assert.NoError(t, err)
+	assert.NotNil(t, res)
+	assert.NotNil(t, res.Body)
+	assert.Empty(t, res.Context)
+
+	assert.Equal(t, "hello", res.String())
+}
+
+func Test_Unix_Start(t *testing.T) {
+	ctx := context.Background()
+	ls, err := net.Listen("unix", "sock.unix")
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.NoError(t, err)
 	assert.NotNil(t, w)
 
@@ -239,63 +328,86 @@ func Test_Unix_Start2(t *testing.T) {
 	}
 }
 
-func Test_Unix_Failboot2(t *testing.T) {
+func Test_Unix_Failboot(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(t, err)
-	}()
+	ctx := context.Background()
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/failboot.php")
 
-	w, err := NewSocketServer(ls, time.Second*5).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Second*5, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err)
 }
 
-func Test_Unix_Timeout2(t *testing.T) {
+func Test_Unix_Timeout(t *testing.T) {
 	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(t, err)
-	}()
+	ctx := context.Background()
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/slow-client.php", "echo", "unix", "200", "0")
 
-	w, err := NewSocketServer(ls, time.Millisecond*100).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Millisecond*100, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Nil(t, w)
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "relay timeout")
+	assert.Contains(t, err.Error(), "context deadline exceeded")
 }
 
-func Test_Unix_Invalid2(t *testing.T) {
+func Test_Unix_Invalid(t *testing.T) {
+	ctx := context.Background()
 	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(t, err)
-	}()
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/invalid.php")
 
-	w, err := NewSocketServer(ls, time.Second*10).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Second*10, log).SpawnWorkerWithTimeout(ctx, cmd)
 	assert.Error(t, err)
 	assert.Nil(t, w)
 }
 
-func Test_Unix_Broken2(t *testing.T) {
+func Test_Unix_Broken(t *testing.T) {
+	ctx := context.Background()
 	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		errC := ls.Close()
-		assert.NoError(t, errC)
-	}()
+	if err == nil {
+		defer func() {
+			errC := ls.Close()
+			if errC != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "broken", "unix")
-
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -312,24 +424,31 @@ func Test_Unix_Broken2(t *testing.T) {
 
 	assert.Error(t, err)
 	assert.Nil(t, res)
-	wg.Wait()
 
 	time.Sleep(time.Second)
 	err = w.Stop()
 	assert.NoError(t, err)
+
+	wg.Wait()
 }
 
-func Test_Unix_Echo2(t *testing.T) {
+func Test_Unix_Echo(t *testing.T) {
+	ctx := context.Background()
 	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(t, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(t, err)
-	}()
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				t.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		t.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -355,19 +474,25 @@ func Test_Unix_Echo2(t *testing.T) {
 	assert.Equal(t, "hello", res.String())
 }
 
-func Benchmark_Tcp_SpawnWorker_Stop2(b *testing.B) {
-	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(b, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(b, err)
-	}()
+func Benchmark_Tcp_SpawnWorker_Stop(b *testing.B) {
+	ctx := context.Background()
+	ls, err := net.Listen("tcp", "127.0.0.1:9007")
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				b.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		b.Skip("socket is busy")
+	}
 
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	for n := 0; n < b.N; n++ {
 		cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-		w, err := f.SpawnWorker(cmd)
+		w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
 		if err != nil {
 			b.Fatal(err)
 		}
@@ -382,17 +507,23 @@ func Benchmark_Tcp_SpawnWorker_Stop2(b *testing.B) {
 	}
 }
 
-func Benchmark_Tcp_Worker_ExecEcho2(b *testing.B) {
-	ls, err := net.Listen("unix", "sock.unix")
-	assert.NoError(b, err)
-	defer func() {
-		err = ls.Close()
-		assert.NoError(b, err)
-	}()
+func Benchmark_Tcp_Worker_ExecEcho(b *testing.B) {
+	ctx := context.Background()
+	ls, err := net.Listen("tcp", "127.0.0.1:9007")
+	if err == nil {
+		defer func() {
+			err = ls.Close()
+			if err != nil {
+				b.Errorf("error closing the listener: error %v", err)
+			}
+		}()
+	} else {
+		b.Skip("socket is busy")
+	}
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "tcp")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -412,15 +543,13 @@ func Benchmark_Tcp_Worker_ExecEcho2(b *testing.B) {
 	}
 }
 
-func Benchmark_Unix_SpawnWorker_Stop2(b *testing.B) {
-	defer func() {
-		_ = syscall.Unlink("sock.unix")
-	}()
+func Benchmark_Unix_SpawnWorker_Stop(b *testing.B) {
+	ctx := context.Background()
 	ls, err := net.Listen("unix", "sock.unix")
 	if err == nil {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
 				b.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -428,11 +557,11 @@ func Benchmark_Unix_SpawnWorker_Stop2(b *testing.B) {
 		b.Skip("socket is busy")
 	}
 
-	f := NewSocketServer(ls, time.Minute)
+	f := NewSocketServer(ls, time.Minute, log)
 	for n := 0; n < b.N; n++ {
 		cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-		w, err := f.SpawnWorker(cmd)
+		w, err := f.SpawnWorkerWithTimeout(ctx, cmd)
 		if err != nil {
 			b.Fatal(err)
 		}
@@ -443,15 +572,13 @@ func Benchmark_Unix_SpawnWorker_Stop2(b *testing.B) {
 	}
 }
 
-func Benchmark_Unix_Worker_ExecEcho2(b *testing.B) {
-	defer func() {
-		_ = syscall.Unlink("sock.unix")
-	}()
+func Benchmark_Unix_Worker_ExecEcho(b *testing.B) {
+	ctx := context.Background()
 	ls, err := net.Listen("unix", "sock.unix")
 	if err == nil {
 		defer func() {
-			errC := ls.Close()
-			if errC != nil {
+			err = ls.Close()
+			if err != nil {
 				b.Errorf("error closing the listener: error %v", err)
 			}
 		}()
@@ -461,7 +588,7 @@ func Benchmark_Unix_Worker_ExecEcho2(b *testing.B) {
 
 	cmd := exec.Command("php", "../../tests/client.php", "echo", "unix")
 
-	w, err := NewSocketServer(ls, time.Minute).SpawnWorker(cmd)
+	w, err := NewSocketServer(ls, time.Minute, log).SpawnWorkerWithTimeout(ctx, cmd)
 	if err != nil {
 		b.Fatal(err)
 	}
